complaint_chatbot/requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0

# LLM API clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Azure OpenAI (optional)
# azure-openai>=1.0.0

# Utilities
requests>=2.31.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
